refactor(desktop): organize auth hooks
