+++ desktop/src/App.tsx
@@ -13,7 +13,7 @@ import { APP_ROUTES, LEGACY_APP_ROUTES } from "@/config/app-routes"
 import { useAppCommandActions } from "@/hooks/app/use-app-command-actions"
 import { useExportRunningAgentCount } from "@/hooks/app/use-export-running-agent-count"
 import { useAppShortcuts } from "@/hooks/app/use-app-shortcuts"
-import { useAuthBootstrap } from "@/hooks/auth/use-auth-bootstrap"
+import { useAuthBootstrap } from "@/hooks/auth/lifecycle/use-auth-bootstrap"
 import { useAgentAutoReconcile } from "@/hooks/agents/lifecycle/use-agent-auto-reconcile"
 import { useLocalAutomationExecutor } from "@/hooks/automations/lifecycle/use-local-automation-executor"
 import { useRuntimeInputSyncRuntime } from "@/hooks/cloud/use-runtime-input-sync-runtime"
