+++ desktop/src/hooks/auth/facade/use-login-page.ts
@@ -1,9 +1,10 @@
 import { useLocation, useNavigate } from "react-router-dom";
-import { useGitHubSignIn } from "@/hooks/auth/use-github-sign-in";
+import { useGitHubSignIn } from "@/hooks/auth/workflows/use-github-sign-in";
 import { isProductAuthRequired } from "@/lib/domain/auth/auth-mode";
 import { useAuthStore } from "@/stores/auth/auth-store";
 import { getRedirectTarget } from "@/lib/domain/auth/login-redirect";
 
+// Owns the login page view model by composing auth state and sign-in actions.
 export function useLoginPage() {
   const navigate = useNavigate();
   const location = useLocation();
